feat: extract InputWithTemplate component
