+++ components/InputWithTemplate/AutoWidthInput.vue
