+++ components/InputWithTemplate/index.ts
@@ -1,0 +1,7 @@
+import InputWithTemplate from "./InputWithTemplate.vue";
+import { Template } from "./types";
+
+export {
+  InputWithTemplate,
+  Template
+}
