+++ components/InputWithTemplate/types.ts
@@ -1,0 +1,4 @@
+export interface Template {
+  id: string;
+  tip?: string;
+}
